Reduce code redundancy by using `RawParser` and `RawParserExt`
Also added changelog
